@@ -5,51 +5,46 @@ namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Auth\Events\Registered;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rules;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\DB;
-use Psy\Readline\Hoa\Console;
-
-
+use Illuminate\View\View;
 
 class RegisteredUserController extends Controller
 {
-    public function store(Request $request)
+    /**
+     * Display the registration view.
+     */
+    public function create(): View
     {
-        \Log::info('Store method called');
+        return view('auth.register');
+    }
 
+    /**
+     * Handle an incoming registration request.
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     */
+    public function store(Request $request): RedirectResponse
+    {
         $request->validate([
-            'username' => ['required', 'string', 'max:255', 'unique:clients,username'],
-            'password' => ['required', 'string', 'min:8', 'confirmed'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:clients,email'],
-            'phoneNumber' => ['required', 'string', 'regex:/^[\d\s+\-()]*$/', 'unique:clients,phoneNumber'],
-            'firstName' => ['required', 'string', 'max:255'],
-            'lastName' => ['required', 'string', 'max:255'],
+            'name' => ['required', 'string', 'max:255'],
+            'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:'.User::class],
+            'password' => ['required', 'confirmed', Rules\Password::defaults()],
         ]);
-        \Log::info('Validation passed');
-        
-        try {
-            $clients = User::create([
-                'username' => $request->username,
-                'email' => $request->email,
-                'password' => Hash::make($request->password),
-                'phoneNumber' => $request->phoneNumber,
-                'firstName' => $request->firstName,
-                'lastName' => $request->lastName,
-            ]);
-        } catch (\Exception $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->withErrors(['msg' => 'User creation failed'])->withInput();
 
-        } 
+        $user = User::create([
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => Hash::make($request->password),
+        ]);
 
-        event(new Registered($clients));
+        event(new Registered($user));
 
-        Auth::login($clients);
+        Auth::login($user);
 
-        return redirect()->intended('dashboard');
+        return redirect(route('dashboard', absolute: false));
     }
 }
